email-classifier-graphrag/requirements.txt
# Core ML/AI
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
scikit-learn>=1.3.0
numpy<2.0.0
pandas>=2.0.0

# Graph & Neo4j
neo4j>=5.0.0
py2neo>=2021.2.3
networkx>=3.0

# Embeddings & RAG
sentence-transformers>=2.2.0
langchain>=0.0.300
faiss-cpu>=1.7.4
openai>=0.28.0

# Web & API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
colorama>=0.4.6

# Development
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
